feat: 타겟 date가 해당하는 기수의 몇주차인지 확인하는 함수 추가
